refactor(theme-editor): modularize preview utilities and enhance theme configurator
- Introduced a new `preview-utils.ts` file to encapsulate shared preview building blocks for better code organization and reusability.
- Re-exported essential utilities from `preview-utils` in the theme-editor for streamlined access.
- Refactored theme configurator to utilize new utilities, improving maintainability and reducing code duplication.
- Updated the theme configurator to remove legacy constants and functions, enhancing clarity and performance.
